+++ src/Assert.php
@@ -38,7 +38,7 @@ class Assert
      * @psalm-pure
      * @psalm-assert string $value
      *
-     * @param mixed  $value
+     * @param mixed $value
      * @param string $message
      *
      * @throws InvalidArgumentException
@@ -57,7 +57,7 @@ class Assert
      * @psalm-pure
      * @psalm-assert non-empty-string $value
      *
-     * @param mixed  $value
+     * @param mixed $value
      * @param string $message
      *
      * @throws InvalidArgumentException
@@ -72,7 +72,7 @@ class Assert
      * @psalm-pure
      * @psalm-assert int $value
      *
-     * @param mixed  $value
+     * @param mixed $value
      * @param string $message
      *
      * @throws InvalidArgumentException
@@ -91,14 +91,14 @@ class Assert
      * @psalm-pure
      * @psalm-assert numeric $value
      *
-     * @param mixed  $value
+     * @param mixed $value
      * @param string $message
      *
      * @throws InvalidArgumentException
      */
     public static function integerish($value, $message = '')
     {
-        if (!\is_numeric($value) || $value != (int) $value) {
+        if (!\is_numeric($value) || $value != (int)$value) {
             static::reportInvalidArgument(\sprintf(
                 $message ?: 'Expected an integerish value. Got: %s',
                 static::typeToString($value)
@@ -110,7 +110,7 @@ class Assert
      * @psalm-pure
      * @psalm-assert positive-int $value
      *
-     * @param mixed  $value
+     * @param mixed $value
      * @param string $message
      *
      * @throws InvalidArgumentException
@@ -120,6 +120,44 @@ class Assert
         if (!(\is_int($value) && $value > 0)) {
             static::reportInvalidArgument(\sprintf(
                 $message ?: 'Expected a positive integer. Got: %s',
+                static::valueToString($value)
+            ));
+        }
+    }
+
+    /**
+     * @psalm-pure
+     * @psalm-assert non-negative-int $value
+     *
+     * @param mixed $value
+     * @param string $message
+     *
+     * @throws InvalidArgumentException
+     */
+    public static function nonNegativeInteger($value, $message = '')
+    {
+        if (!(\is_int($value) && $value >= 0)) {
+            static::reportInvalidArgument(\sprintf(
+                $message ?: 'Expected a non negative integer. Got: %s',
+                static::valueToString($value)
+            ));
+        }
+    }
+
+    /**
+     * @psalm-pure
+     * @psalm-assert non-negative-int $value
+     *
+     * @param mixed $value
+     * @param string $message
+     *
+     * @throws InvalidArgumentException
+     */
+    public static function negativeInteger($value, $message = '')
+    {
+        if (!(\is_int($value) && $value < 0)) {
+            static::reportInvalidArgument(\sprintf(
+                $message ?: 'Expected a negative integer. Got: %s',
                 static::valueToString($value)
             ));
         }